BugFixes + feature testing
Changed a lot of types in parameters and returns of many methods. Fixed a lot of problems due to bad implementations. Took care of the min and max value in the tree (the algorithm behind their calculations). Added functionalities to let the code be more smooth to write. Started testing and debugging.